# openapi/external.yaml
openapi: 3.1.0
info:
  title: agent-orc External Task API
  description: |
    Programmable interface for submitting tasks to agent-orc from external systems.
    Each task creates an `AgentRun` in the caller's tenant namespace and returns a
    task id that can be polled, streamed, answered (for clarification), or cancelled.

    Authentication: every endpoint except `/oauth/token` requires
    `Authorization: Bearer <token>`, where the token is either an agent-orc-issued
    JWT (from the `client_credentials` exchange below), a federated OIDC JWT, or a
    Kubernetes ServiceAccount token. See docs/auth.md for the full model.
  version: "1.0.0"
  contact:
    name: agent-orc
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8084
    description: local dev / cluster-internal
  - url: https://agent-orc.example.com:8084
    description: production (exposed via Ingress/Gateway + UIProxy BFF)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT-or-K8s-SA-token
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
    TokenRequest:
      description: form-encoded body for the client_credentials grant.
      type: object
      properties:
        grant_type:
          type: string
          enum: [client_credentials]
          description: Must be `client_credentials`.
        client_id:
          type: string
        client_secret:
          type: string
    TokenResponse:
      type: object
      required: [access_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: Seconds until expiry (default 3600).
...
paths:
  /oauth/token:
    post:
      tags: [authentication]
      summary: Exchange client credentials for a bearer token
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '200':
          description: Access token issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Unsupported grant type or malformed body.
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '401':
          description: Invalid client id or secret.
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...